functionality should be done
Still needs testing
Still needs an actual design (colours etc)
