+++ mjako18/assignment1/gallery.php
@@ -30,16 +30,24 @@
 </head>
 <body>
   <header>
-    <nav class="nav navbar-fixed">
-      <div class="container">
-        <ul>
-          <li class="menu skew-right float_right skew-right-end<?php echo $menu_logout ?>"><a href="uploadform.php">upload</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_logout ?>"><a href="gallery.php">pictures</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_logout ?>"><a href="users.php">users</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_login ?>"><a href="register_form.php">register</a></li>
-          <li class="menu skew-right float_right"><a href="index.php">frontpage</a></li>
-        </ul>
-      </div>
+    <nav class="nav navbar-fixed hidden">
+      <ul class="hide-s">
+        <li class="menu skew-right float_right menu-right-end hidden"><a href="uploadform.php">upload</a></li>
+        <li class="menu skew-right float_right hidden menu-current"><a href="gallery.php">pictures</a></li>
+        <li class="menu skew-right float_right hidden"><a href="users.php">users</a></li>
+        <li class="menu skew-right float_right shown menu-right-end"><a href="register_form.php">register</a></li>
+        <li class="menu skew-right float_right"><a href="index.php">frontpage</a></li>
+      </ul>
+<!-- folding menu -->
+      <div class="hide-l hide-m menu-toggle" onclick="menu_toggle();"><span class="float_right">&#9776;</span></div>
+      <ul id="mobile-menu" class="hide-l hide-m hide-s">
+        <li class="menu"><a href="index.php">frontpage</a></li>
+        <li class="menu shown"><a href="register_form.php">register</a></li>
+        <li class="menu hidden"><a href="users.php">users</a></li>
+        <li class="menu hidden menu-current"><a href="gallery.php">pictures</a></li>
+        <li class="menu hidden"><a href="uploadform.php">upload</a></li>
+<!-- folding menu -->
+      </ul>
     </nav>
   </header>
   <section>
@@ -50,13 +58,36 @@
         <div id="logoutDiv" class="<?php echo $logout_classes ?>">
           <button name="logout" id="logout" class="button button-logout" onclick="logout('gallery');">Log out</button>
         </div>
+        <div id="loginFormDiv" class="<?php echo $login_classes ?>">
+          <form action="login.php" method="post" id="loginform">
+            <fieldset>
+              <legend>Login</legend>
+              <div>
+                <label for="username">Username:</label>
+                <input type="text" name="username" id="username" class="" placeholder="username">
+              </div>
+              <div>
+                <label for="pass">Password:</label>
+                <input type="password" name="pass" id="pass" class="" placeholder="password">
+              </div>
+              <div>
+                <button name="login" id="login" class="button button-login" form="loginform" onclick="return postLogin(this.form);">Log in</button> | <button name="" id="" class="button button-register" "register.php">Register account</button>
+              </div>
+            </fieldset>
+          </form>
+        </div>
+        <div id="login_response"></div>
       </div>
     </div>
   </section>
 
   <footer>
   </footer>
-  <script type="text/javascript" src="js/ajax.js"></script>
+  <script type="text/javascript" src="./js/util.js"></script>
+  <script type="text/javascript" src="./js/menu.js"></script>
+  <script type="text/javascript" src="./js/login.js"></script>
+  <script>checkLoggedIn();</script>
+  <script type="text/javascript" src="js/gallery.js"></script>
   <script>
     document.addEventListener('DOMContentLoaded', getImages, false);
   </script>
