+++ mjako18/assignment1/index.php
@@ -15,28 +15,36 @@
     $logout_classes = " hidden";
     $login_classes = "";
     $menu_logout = " hidden";
-    $menu_login = " shown skew-right-end";
+    $menu_login = " shown menu-right-end";
   }
 ?>
 <!DOCTYPE html>
 <html>
 <head>
   <title>Frontpage</title>
-  <link type="text/css" rel="stylesheet" href="CSS/responsive.css">
-  <link type="text/css" rel="stylesheet" href="CSS/design.css">
+  <link type="text/css" rel="stylesheet" href="./CSS/responsive.css">
+  <link type="text/css" rel="stylesheet" href="./CSS/design.css">
 </head>
 <body>
   <header>
     <nav class="nav navbar-fixed">
-      <div class="container">
-        <ul>
-          <li class="menu skew-right float_right skew-right-end<?php echo $menu_logout ?>"><a href="uploadform.php">upload</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_logout ?>"><a href="gallery.php">pictures</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_logout ?>"><a href="users.php">users</a></li>
-          <li class="menu skew-right float_right<?php echo $menu_login ?>"><a href="register_form.php">register</a></li>
-          <li class="menu skew-right float_right"><a href="index.php">frontpage</a></li>
-        </ul>
-      </div>
+      <ul class="hide-s">
+        <li class="menu skew-right float_right menu-right-end hidden"><a href="uploadform.php">upload</a></li>
+        <li class="menu skew-right float_right hidden"><a href="gallery.php">pictures</a></li>
+        <li class="menu skew-right float_right hidden"><a href="users.php">users</a></li>
+        <li class="menu skew-right float_right shown menu-right-end"><a href="register_form.php">register</a></li>
+        <li class="menu skew-right float_right menu-current"><a href="index.php">frontpage</a></li>
+      </ul>
+<!-- folding menu -->
+      <div class="hide-l hide-m menu-toggle" onclick="menu_toggle();"><span class="float_right">&#9776;</span></div>
+      <ul id="mobile-menu" class="hide-l hide-m hide-s">
+        <li class="menu menu-current"><a href="index.php">frontpage</a></li>
+        <li class="menu shown"><a href="register_form.php">register</a></li>
+        <li class="menu hidden"><a href="users.php">users</a></li>
+        <li class="menu hidden"><a href="gallery.php">pictures</a></li>
+        <li class="menu hidden"><a href="uploadform.php">upload</a></li>
+<!-- folding menu -->
+      </ul>
     </nav>
   </header>
   <section>
@@ -47,10 +55,10 @@
         </div>
       </div>
       <div class="col-xl-4 col-m-4">
-        <div id="logoutDiv" class="<?php echo $logout_classes ?>">
+        <div id="logoutDiv" class="hidden">
           <button name="logout" id="logout" class="button button-logout" onclick="logout('index');">Log out</button>
         </div>
-        <div id="loginFormDiv" class="<?php echo $login_classes ?>">
+        <div id="loginFormDiv" class="shown">
           <form action="login.php" method="post" id="loginform">
             <fieldset>
               <legend>Login</legend>
@@ -63,7 +71,7 @@
                 <input type="password" name="pass" id="pass" class="" placeholder="password">
               </div>
               <div>
-                <button name="login" id="login" class="button button-login" form="loginform" onclick="return postForm(this.form, 'login');">Log in</button> | <button name="" id="" class="button button-register" "register.php">Register account</button>
+                <button name="login" id="login" class="button button-login" form="loginform" onclick="return postLogin(this.form);">Log in</button> | <button name="" id="" class="button button-register" "register.php">Register account</button>
               </div>
             </fieldset>
           </form>
@@ -75,6 +83,9 @@
 
   <footer>
   </footer>
-  <script type="text/javascript" src="js/ajax.js"></script>
+  <script type="text/javascript" src="./js/util.js"></script>
+  <script type="text/javascript" src="./js/login.js"></script>
+  <script type="text/javascript" src="./js/menu.js"></script>
+  <script>checkLoggedIn();</script>
 </body>
 </html>
